SOF-263: Refactor error handling files and clamp bitmap cropping dimensions
