+++ app/src/main/java/com/vci/vectorcamapp/core/presentation/components/BaseScaffold.kt
@@ -4,17 +4,15 @@ import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.padding
+import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Scaffold
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.CompositionLocalProvider
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.unit.dp
-import com.vci.vectorcamapp.core.presentation.util.error.ErrorData
 import com.vci.vectorcamapp.core.presentation.util.error.ErrorMessageBus
-import com.vci.vectorcamapp.core.presentation.util.error.ErrorSnackbarHost
 import com.vci.vectorcamapp.core.presentation.util.error.LocalErrorDataFlow
-import kotlinx.coroutines.flow.SharedFlow
+import com.vci.vectorcamapp.ui.extensions.dimensions
 
 @Composable
 fun BaseScaffold(
@@ -30,7 +28,7 @@ fun BaseScaffold(
                 ErrorSnackbarHost(
                     modifier = Modifier
                         .align(Alignment.TopCenter)
-                        .padding(top = 16.dp)
+                        .padding(top = MaterialTheme.dimensions.paddingMedium)
                 )
             }
         }
