+++ app/src/main/java/com/vci/vectorcamapp/core/presentation/util/error/ErrorMessageBus.kt
@@ -2,6 +2,7 @@ package com.vci.vectorcamapp.core.presentation.util.error
 
 import androidx.compose.material3.SnackbarDuration
 import com.vci.vectorcamapp.core.domain.util.Error
+import com.vci.vectorcamapp.core.presentation.model.ErrorData
 import kotlinx.coroutines.channels.BufferOverflow
 import kotlinx.coroutines.flow.MutableSharedFlow
 import kotlinx.coroutines.flow.asSharedFlow
